refactor(kill_server): 重构插件入口点和事件定义
- 将事件定义移至独立的 events.py 文件
- 创建新的 entry.py 作为插件入口点文件
- 从 __init__.py 中移除所有实现代码
- 更新插件版本号至 1.0.1-beta.1
- 添加插件入口点配置
- 优化全局变量引用声明
- 改进 on_info 事件处理器以跳过玩家消息
