+++ src/KillServer/mcdreforged.plugin.json
@@ -1,6 +1,6 @@
 {
     "id": "kill_server",
-    "version": "1.0.0",
+    "version": "1.0.1-beta.1",
     "name": "KillServer",
     "description": {
         "en_us": "Force to kill mc server",
@@ -12,5 +12,6 @@
     "dependencies": {
         "python": ">=3.12",
         "mcdreforged": ">=2.15"
-    }
+    },
+    "entrypoint": "kill_server.entry"
 }
